Initial Commit: Calico ver 0.4.1 pre-alpha
This is the first commit for my voice assistant project that relies on Rhasspy 2.5.11 for most of its functionality. Design and testing uses a VirtualBox Ubuntu 24.04.2 LTS vm with guest additions and the Rhasspy 2.511 Docker (CLI version) container. Calico sits directly on Ubuntnu where it communicates with Rhasspy via the included MQTT broker. The sentences.ini and profile.json files are crucial for proper setup and must replace the default versions indluded in Rhasspy's folder under .config. README coming next.