Added installer, will be part of v0.5.0 pre-alpha
Intended to work on any Debian platform, robustly installs any and all potential dependencies. Make sure OS is set to install "All Updates" and not just security updates. This last step is REQUIRED.

Going to build the launcher next. The UI will use PyQT6. Settings will be a part of the launcher and no longer a stand-alone application.

Also planning to revise README.