Calico v0.5.0 pre-alpha
Updated to use partially pre-configured rhasspy config directory. This enables a rough version of the "hey calico" wake word, but also shows users where to go to learn how to change this feature. The script also tells users to go to the Rhasspy web UI to download required files on first launch.